Basic snake movement.
No tail or edge detection yet - just moving snake around based on its
'position' - a byte 'pos' value which can be converted into an X and Y
co-ord.

This is so that we can easily store the full snake in a byte-array ring
buffer when we come to adding segments to it.