Use a ring buffer to have a snake with length
The ring buffer 'snakeBuffer' holds the 'squares' of the snake
(where each square is a distinct location within the ROOM).

The principle is that we maintain two pointers to the snake in this
buffer -- the head and the tail. Every location in between them is a place
where the snake head 'has been', and is as such part of the body.

Maintaining the 'tail' location is a matter of moving the tail position pointer
forwards with the head pointer, or -- if we want to make the snake longer -- leaving
it as-is.

Added a random function to say whether we extend the snake, until fruit munch is added.

Also added a limitation to say what the maximum size of the snake should be. This has a hard
limit of SNAKE_BUFFER_SIZE - 1, so that do not end up with the head position catching up with the
tail.