Clarify the logic around collision-with-edge
The collision-with-edge is a special case for collisions, since we
are establishing that we *will* collide with a thing rather than
that we *have* collided -- we calculate based on the current position
rather than the location we will move to.

Splitting all this out into its own function allows us to see the logic
more clearly and treat it as the edge-case it is.