Add Fruit dynamic to game.
A new fruit will appear at a random position as soon as the current
fruit has been eaten. This is how the snake grows and how we score
points.

The way we're chosing the fruit location at the moment is tremendously
inefficient - it randomly picks a square and sees if that is 'within the
snake' -- and if it is then picks another, ad infinitum via recursion. This
will get daft at really long snake length, but seems to work ok with the snake
up to 62 squares.